Extract tensor validation helpers and narrow exception scope in JEPA-WMS provider
Move the _is_sequence/_shape/_require_rank/_flatten_numeric/_tensor_shape
helpers out of providers/jepa_wms.py into a new private sibling module
providers/_tensor_validation.py so the JEPA-WMS provider file focuses on
the adapter contract rather than shape plumbing.

Narrow the broad 'except Exception' catches around torch.hub.load, as_tensor,
encode/unroll/normalize, action-tensor preparation, timestep slicing, and the
outer score_actions safety net to the specific exception families those call
sites can raise (AttributeError/TypeError/ValueError/RuntimeError/ImportError/
OSError and the indexing set), so unexpected failures propagate instead of
being silently wrapped as ProviderError.

Public classes, methods, capability flags, registration status, and all
existing ProviderError message strings are unchanged.
